refactor: move sorting ops out of lax_numpy
